refactor: moved viewmodels to the correct folder
