+++ app/src/main/java/com/example/sporthub/ui/venueList/VenueListFragment.kt
@@ -15,13 +15,12 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.core.app.ActivityCompat
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
-import androidx.lifecycle.Observer
 import androidx.navigation.fragment.findNavController
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.example.sporthub.R
 import com.example.sporthub.data.model.Venue
-import com.example.sporthub.ui.findVenues.FindVenuesViewModel
+import com.example.sporthub.viewmodel.FindVenuesViewModel
 import com.example.sporthub.ui.venueList.adapter.VenueAdapter
 import com.google.android.gms.location.FusedLocationProviderClient
 import com.google.android.gms.location.LocationServices
